refinement implementation finished for lbe initial testing showed system working so far needs more testing and needs to identify efficiency issues
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@665 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/cpa/symbpredabs/mathsat/MathsatSymbolicFormulaManager.java
@@ -887,7 +887,7 @@ public class MathsatSymbolicFormulaManager implements SymbolicFormulaManager {
             } catch (NumberFormatException nfe) {
                 long l = Long.parseLong(num);
                 if (l < 0) {
-                    num = Long.toString((long)Integer.MAX_VALUE + l);
+                    num = Long.toString(Integer.MAX_VALUE + l);
                 } else {
                     num = Long.toString(l - ((long)Integer.MAX_VALUE + 1)*2);
                 }
